fix typings in extras_provider.py
